+++ og-object/src/main/java/com/ibm/og/object/WriteCopyObjectNameConsumer.java
@@ -69,7 +69,7 @@ public class WriteCopyObjectNameConsumer extends AbstractObjectNameConsumer {
     final long objectSize = getSourceObjectSize(request);
     final int containerSuffix = getSourceObjectContainerSuffix(request);
     final byte numLegalHolds = getNumberOfLegalHolds(request, response);
-    final long retention = getObjectionRetention(request);
+    final int retention = getObjectionRetention(request);
     if (objectString == null) {
       return null;
     } else {
